Add files via upload
This is the most up to date version of the Phone book assignment for CS 145 at WCC. It has functioning new entry and search --> print methods that are called from a do while loop based menu. The Object class has remained unchanged. Currently the edit function is in development and will use the index architecture that is built into the search method. once edit is complete next steps will be to build the delete feature which should be based in the edit methodology.