Add files via upload
I missed a few uploads and thus there are some gaps in my progress on this repository. However I want to note that the earliest uploads should be considered obsolete. I was initially confused about the object class and how to initialize it. The files contained in this folder are a quasi-spike exercise. I am glad to have stepped back and worked on a very basic version of a linked list in OOP as it now serves as the backbone of the project. Thus I feel it is important to include them in this repository.