Add files via upload
The most up to date (and possibly complete) phone book yet! 
1. delete method has been created and is function
2. can now print the whole phone book via the addition of a new menu option
3. updated header entry

Current thoughts:
This program appears to be pretty robust against miss-inputs and can work with a different number of entries. It has some default behaviors that seem bad at first but appear to prevent crashes and allow the user to try agian (See search method while loop). The user interface is number based which is simple but easy to use. Print formatting is adequate. The biggest downside is that all methods start at the front of the linked list and traverse it till the given index every time, which is not super efficient. A thing I may fix is to extract the loop in the cases 1,2,&3 to a method. I may also leave it as is for clarity purposes though.